build-script: add zImage check

+++ build.sh
@@ -3,7 +3,10 @@ TOOLCHAIN="/home/varunchitre15/toolchains/arm-eabi-linaro-4.6.2/bin/arm-eabi"
 MODULES_DIR="../modules"
 ZIMAGE="/home/varunchitre15/kernel/kernel/arch/arm/boot/zImage"
 KERNEL_DIR="/home/varunchitre15/kernel/kernel"
+if [ -a $KERNEL_DIR/arch/arm/boot/zImage ];
+then
 rm $ZIMAGE
+fi
 make ARCH=arm CROSS_COMPILE=$TOOLCHAIN- sa77_defconfig
 make ARCH=arm CROSS_COMPILE=$TOOLCHAIN- -j8
 if [ -a $KERNEL_DIR/arch/arm/boot/zImage ];
